% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BFreplicate}
\alias{BFreplicate}
\title{Bayes Factors Using Bayesm}
\usage{
BFreplicate(sim, n = 100, numRep = 20, iters = 2000)
}
\arguments{
\item{mydata,}{alpha, Krange}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
# see bayesm documentation
sim1_n100<-sim1func(n=300)
BF<-getBF(sim1_n100$Y, 1, c(1:6))
BF
}
\keyword{bayes}
\keyword{bayesm}
\keyword{factor,}

