% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCSV4Plot.r
\name{MakeCSV4Plot}
\alias{MakeCSV4Plot}
\title{This is MakeCSV4Plot}
\usage{
MakeCSV4Plot(.df, FileName = "test", PlotNames = OfficialNames)
}
\arguments{
\item{stuff}{and more}
}
\description{
This is MakeCSV4Plot
}
\examples{
#not run
}
\keyword{multi}
