% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maxZ.r
\name{maxZ}
\alias{maxZ}
\title{maxZ}
\usage{
maxZ(x)
}
\arguments{
\item{mydata,}{alpha, Krange}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
# see bayesm documentation
#sim1_n100<-sim1func(n=300)
maxZ...
maxZ
}
\keyword{bayes}
\keyword{bayesm}
\keyword{factor,}

