% Generated by roxygen2 (4.0.1): do not edit by hand
\name{permutations}
\alias{permutations}
\title{All permutations given size}
\usage{
permutations(n)
}
\arguments{
\item{n,}{number of values in permutation}
}
\description{
This function outputs all permutations of a set number of values
}
\examples{
permutations(2)
}
\keyword{permutationa}

