% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PostPredFunk}
\alias{PostPredFunk}
\title{PostProcessing Predictive function}
\usage{
PostPredFunk(.GrunK0us = GrunK0us, .Zetc = Zetc, .Y = Y)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

