% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PostPredFunkMVN}
\alias{PostPredFunkMVN}
\title{PostProcessing Predictive function}
\usage{
PostPredFunkMVN(.GrunK0us, .Zetc, .Y, .prep, .simlabel)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

