% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PostProc_CSIRO}
\alias{PostProc_CSIRO}
\title{PostProcessing function univ}
\usage{
PostProc_CSIRO(Grun, mydata, LineUp = 1, prep = 10000, Propmin = 0.01,
  isSim = FALSE, simlabel = "sim", savelabel = "PPplot", nEnd = 2000,
  PlotType = "Boxplot")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

