% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PostProc_mvn}
\alias{PostProc_mvn}
\title{PostProcessing function univ}
\usage{
PostProc_mvn(Grun, mydata, Propmin = 0.05, isSim = TRUE, simlabel = "sim",
  savelabel = "PPplot")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

