% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProc_mvn_Oct15.r
\name{PostProc_mvn_Oct15}
\alias{PostProc_mvn_Oct15}
\title{PostProcessing function univ}
\usage{
PostProc_mvn_Oct15(Grun, mydata, Propmin = 0.05, isSim = FALSE,
  simlabel = "sim", savelabel = "PPplot")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}
