% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PostProcUnivariate}
\alias{PostProcUnivariate}
\title{PostProcessing function univ}
\usage{
PostProcUnivariate(Grun, mydata, prep = 10000, LineUp = 1, Propmin = 0.05,
  isSim = TRUE, simlabel = "sim")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

