% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PostProcUnivariate_later}
\alias{PostProcUnivariate_later}
\title{PostProcessing function univ}
\usage{
PostProcUnivariate_later(Grun, mydata, LineUp = 1, prep = 10000,
  Propmin = 0.05, isSim = TRUE, simlabel = "sim", savelabel = "PPplot")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

