% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcUnivariate_later2.r
\name{PostProcUnivariate_later2}
\alias{PostProcUnivariate_later2}
\title{PostProcessing function univ}
\usage{
PostProcUnivariate_later2(Grun, mydata, LineUp = 1, prep = 10000,
  Propmin = 0.05, isSim = TRUE, simlabel = "sim", savelabel = "PPplot",
  nEnd = 2000)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}
