% Generated by roxygen2 (4.0.1): do not edit by hand
\name{QuickSwitch_allPars}
\alias{QuickSwitch_allPars}
\title{Label switch fix for Univariate Normals}
\usage{
QuickSwitch_allPars(GibbResult, isSim = TRUE, LineUpBy = 1, PropMin = 0.3)
}
\arguments{
\item{GibbResult:}{output of Gibbs run in a list

        [[Mu]] matrix of mu samples  (rows=iterations, columns  = components),

        [[Sig]] matrix variance  (really Sig^2, but doesnt matter here anyway)

        [Ps]] matrix weights

        [[Loglike]] Loglikelihood (vector length of iterations)

        [[SteadyScore]] vector of number of non-empty groups at each iteration  (SET to NULL if not computed, not important here)

        [[Zs]] estimated component (allocation) for each Y at each iteration (rows=n by columns=Iterations)

        [[Y]] for simulations, list of Y$Y and Y$Z, the observed data and true groups. set Y$Z to NULL if unknown.}

\item{LineUpBy}{: currently can be set to 1 or 2. It is only used to rename the FINAL output
        1: final output will order Components by posterior weight (Z=1 for max(weight),etc)
        2: final output ordered by variance of components
    Again, note this doesnt affect the actualy unswitching, which uses no ordering. it simply relabels the reference groups according to chosen parameter to nice output.}

\item{PropMin}{= Minimum allowable proportion of a group required to activate second level check.,
    Usually 0.3 sufficient.
    Smaller values increase time function takes, but are needed for Smaller samples or very close components (try 0.1)
If function gives error, this is first thing to try (make it smaller)}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#sorry, none yet
# requires a constant number of non-empty groups in samples,
# but this can be smaller than the number of components fitted.
# if you have an unsteady number of posterior  non-empty groups:
#subset posterior samples first by this then unswitch.
}
\keyword{gaussian}
\keyword{gibbs}
\keyword{label}
\keyword{switching}
\keyword{univariate}

