% Generated by roxygen2 (4.0.2): do not edit by hand
\name{QuickSwitchMVN}
\alias{QuickSwitchMVN}
\title{Label switch fix for Multivariate Normals}
\usage{
QuickSwitchMVN(MVNResult, PropMin = 0.01)
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#not run
}
\keyword{Wishart}

