% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleSwitchMVN.R
\name{SimpleSwitchMVN}
\alias{SimpleSwitchMVN}
\title{This function fixes the label switching in the output of a Multivariate Gibbs sampler. Requires a constant number of non-empty groups in samples, but can deal with empty groups.}
\usage{
SimpleSwitchMVN(MVNResult, pickRef = c("random", "last", "MaxLike"),
  NumDim = 2, PropMin = 0.01)
}
\arguments{
\item{pickRef;}{'random'= reference picked as a random iteration
'last'= Reference is final iteration
"MaxLike"= Reference is chosen as iteration maximising Log LIkelihood,  (MVNResult$Loglike)}

\item{MVNResult:}{LIST of output of Gibbs sampler with the following components:

        [[Zs]] estimated component (allocation) for each Y at each iteration (by rows=Iterations, Columns=1:n)  

        [P]] matrix of weights, columns - K, rows = Iterations

        [[Mu]] Data frame  with columns: Mu_1. Mu_2, .. ...,Mu_r ,  k (this is the group), Iteration.  

        [[Loglike]] OPTIONAL: Loglikelihood (vector the length of iterations). Needed only if PickRef=="MaxLike"}
}
\description{
This function fixes the label switching in the output of a Multivariate Gibbs sampler. Requires a constant number of non-empty groups in samples, but can deal with empty groups.
}
\examples{
RESULT$
	Unswitched<-SimpleSwitchMVN( RESULT, pickRef="last")
}
\keyword{Wishart}
