% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simudZ.R
\name{simudZ}
\alias{simudZ}
\title{Simulate Univariate normal Mixture}
\usage{
simudZ(n, mu, sig, p, k)
}
\arguments{
\item{n}{The number of samples to draw}

\item{mu}{The mean}

\item{sig}{the covariance}

\item{p}{the probability weights}

\item{k}{labels to sample from}
}
\description{
Simulate Univariate normal Mixture
}
\examples{
#not run
}
\keyword{Wishart}
