% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Zmix_CSIRO.r
\name{Zmix_CSIRO}
\alias{Zmix_CSIRO}
\title{Run Gibbs sampler with prior tempering}
\usage{
Zmix_CSIRO(Y, k = 10, iter = 5000, LineUp = 2, SaveFileName = "zmix",
  Yname = "Variable", DATA_ADDITIONAL, SaveFileSAME = "ALLRESULTS")
}
\arguments{
\item{y,}{k,iter, isSim=TRUE, alphas=}
}
\description{
This function ...
}
\examples{
#... you know...
}
\keyword{Gibbs}
\keyword{estimation}
\keyword{gaussian,}
\keyword{mixture,}
\keyword{normal,}
\keyword{order}
\keyword{parallel}
\keyword{sampler,}
\keyword{tempering,}
\keyword{univariate,}

