% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Zmix_multi_CSIRO.r
\name{Zmix_multi_CSIRO}
\alias{Zmix_multi_CSIRO}
\title{This functionZmix_multi_tempered}
\usage{
Zmix_multi_CSIRO(YZ, iter, k, alphas, sim = FALSE, Propmin = 0.05,
  simlabel = "sim", savelabel = "PPplot", DATA_ADDITIONAL)
}
\arguments{
\item{stuff}{and more}
}
\description{
This functionZmix_multi_tempered
}
\examples{
#not run
}
\keyword{multi}

