% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zmix_multi_tempered.R
\name{Zmix_multi_tempered}
\alias{Zmix_multi_tempered}
\title{This functionZmix_multi_tempered}
\usage{
Zmix_multi_tempered(YZ, iter, k, alphas, sim = TRUE, EndSize = 500,
  plotting = FALSE)
}
\arguments{
\item{YZ}{a data.frame where each row is the principal component loadings from spikes}

\item{iter}{the total number of MCMC iterations to draw}

\item{k}{the number of mixture components to fit}

\item{alphas}{the indices to use in parallel tempering}

\item{sim}{a flag for telling the function whether the data are in a list or not}

\item{EndSize}{the number of samples at the end of the chain to be kept}

\item{plotting}{do we wanto make progressive plots?}
}
\description{
This functionZmix_multi_tempered
}
\examples{
#not run

}
\keyword{multi}
