% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PostProc_CSIRO.R
\name{PostProc_CSIRO}
\alias{PostProc_CSIRO}
\title{PostProcessing function univ}
\usage{
PostProc_CSIRO(Grun, mydata, LineUp = 1, prep = 10000, Propmin = 0.05,
  isSim = TRUE, simlabel = "sim", savelabel = "PPplot", nEnd = 2000,
  PlotType = "Boxplot")
}
\arguments{
\item{v}{and s}
}
\description{
This function draws samples from a Wishart dist
}
\examples{
#nope
}
\keyword{Wishart}

