% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Zmix_lightLYRA}
\alias{Zmix_lightLYRA}
\title{Run Gibbs sampler with prior tempering}
\usage{
Zmix_lightLYRA(y, k, iter = 20000, isSim = TRUE, burn = 5000,
  alphas = c(30, 20, 10, 5, 3, 1, 0.5, 1/2^(c(2, 3, 4, 5, 6, 7, 8, 10, 12, 15,
  20, 30))))
}
\arguments{
\item{y,}{k,iter, isSim=TRUE, alphas=}
}
\description{
This function ...
}
\examples{
#... you know...
}
\keyword{Gibbs}
\keyword{estimation}
\keyword{gaussian,}
\keyword{mixture,}
\keyword{normal,}
\keyword{order}
\keyword{parallel}
\keyword{sampler,}
\keyword{tempering,}
\keyword{univariate,}

