% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Zmix_multi_tempered}
\alias{Zmix_multi_tempered}
\title{This functionZmix_multi_tempered}
\usage{
Zmix_multi_tempered(YZ, iter, k, alphas, sim = TRUE, EndSize = 500)

Zmix_multi_tempered(YZ, iter, k, alphas, sim = TRUE, EndSize = 500)
}
\arguments{
\item{stuff}{and more}

\item{stuff}{and more}
}
\description{
This functionZmix_multi_tempered

This functionZmix_multi_tempered
}
\examples{
#not run
#not run
}
\keyword{multi}

