% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zmix_multi_tempered_old.R
\name{Zmix_multi_tempered_old}
\alias{Zmix_multi_tempered_old}
\title{This functionZmix_multi_tempered}
\usage{
Zmix_multi_tempered_old(YZ, iter, k, alphas, sim = TRUE, EndSize = 500)
}
\arguments{
\item{stuff}{and more}
}
\description{
This functionZmix_multi_tempered
}
\examples{
#not run
}
\keyword{multi}
